+++ src/Bartlett/CompatInfo/Api/Analyser.php
@@ -33,11 +33,12 @@ class Analyser extends BaseApi
      * Analyse a data source and display results.
      *
      * @param string $source Path to the data source
-     * @param bool   $stop_on_failure Stop execution upon first error generated during lexing, parsing or some other operation
+     * @param array $exclude Sets excluded file or directory names for scanning
+     * @param bool $stop_on_failure Stop execution upon first error generated during lexing, parsing or some other operation
      *
      * @return Profile
      */
-    public function run(string $source, bool $stop_on_failure = null): Profile
+    public function run(string $source, array $exclude = [], bool $stop_on_failure = null): Profile
     {
         return $this->request('analyser/run', 'POST', func_get_args());
     }